Refactor to two classes to support reporting on more tools and having Zoom separated out.
Fixes #49
