+++ vocab.py
@@ -3,17 +3,16 @@ from __future__ import annotations
 # standard libraries
 import time
 from datetime import datetime
-from enum import auto, Enum, IntEnum
+from enum import auto, Enum
 from typing import Dict, Union
 
 import pytz
 
 
+
+
+
 # Enum(s)
-class PlacementType(IntEnum):
-    ZOOM = 0
-
-
 class ValidJobName(Enum):
     """
     Each job name should be defined in ValidJobName.
@@ -22,7 +21,7 @@ class ValidJobName(Enum):
 
     COURSE_INVENTORY = 'course_inventory.inventory.run_course_inventory'
     MIVIDEO = 'mivideo.mivideo_extract.main'
-    CANVAS_ZOOM_MEETINGS = 'online_meetings.canvas_zoom_meetings.main'
+    CANVAS_ZOOM_MEETINGS = 'lti_placements.canvas_zoom_meetings.main'
 
 
 class ValidDataSourceName(Enum):
